Adding cards (#3)
## Summary
This PR implements the cards fields.

Unlike the Namelist fields, these are not extracted from the `.def` files but are instead implemented by hand. This is because the syntax of `card` blocks in `.def` files is quite irregular and cleaner pydantic classes can be written by hand.

### Drawbacks
Using prebuilt models comes with the disadvantage that we do not have version-dependent card fields. In this sense, perhaps it would be worthwhile reconsidering if we really want to support all versions of QE that have a tag.